fix: Pass authOptions to getServerSession for proper session handling
Extract shared authOptions from NextAuth route to app/lib/auth/auth-options.ts
and pass it to all getServerSession calls. This ensures the JWT callbacks
run properly and session.user.id is available in API routes.
